requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
